+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Add support for dynamic rules
 
 ## [0.5.0] - 2018-11-02